deleted Rentals.cs from Logic (dead code), deleted CarRentalData (code moved to Seed.cs)
